requirements.txt
mediapipe==0.10.14
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.0.0
